python_training/requirements.txt
# PyTorch with CUDA 12.1 support (for NVIDIA GPUs)
# If you have CUDA 11.8, use: torch>=2.0.0+cu118
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Core dependencies
numpy>=1.24.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# RL libraries with GPU support
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# Hyperparameter Optimization
optuna>=3.0.0

# Distributed Training (optional)
ray[default]>=2.0.0

# Utilities
scikit-learn>=1.3.0
pandas>=2.0.0
yfinance>=0.2.28

# Optional: Monitor GPU usage
gpustat>=1.1.0
